% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnapATACUtils.R
\name{addMatToSnap}
\alias{addMatToSnap}
\title{Add bmat or gmat to snap}
\usage{
addMatToSnap(
  snap,
  cellGroup = NULL,
  group = NULL,
  outDir = NULL,
  matType = "bmat",
  gencode = "vM16",
  ndp = -1,
  ncores = 2,
  force = TRUE,
  binSize = 5000,
  coverFile = TRUE,
  prefix = "snap"
)
}
\arguments{
\item{snap}{snap object}

\item{cellGroup}{vector of characters
cluster labels for snap default is NULL}

\item{group}{characters which cluster needes
default is NULL}

\item{outDir}{characters default is NULL}

\item{matType}{characters default is "bmat"}

\item{gencode}{characters default is "vM16"}

\item{ndp}{integer default is -1
if its value is larger than 0, then down sampling will be used.}

\item{ncores}{integer default is 2}

\item{force}{bool default is TRUE
if add mat even the mat already exists}

\item{binSize}{integer default is 5000}

\item{coverFile}{bool default is TRUE
when set it TRUE, it will cover the output file if exists.}

\item{prefix}{characters default is "snap", which is used for output.}
}
\value{
snap object with corresponding mat.
}
\description{
Add bmat or gmat to snap
}
