% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{AnnoyBuildIndex}
\alias{AnnoyBuildIndex}
\title{Build the annoy index
Ref: Seurat}
\usage{
AnnoyBuildIndex(data, metric = "euclidean", n.trees = 50)
}
\arguments{
\item{data}{Data to build the index with}

\item{metric}{Distance metric; can be one of "euclidean", "cosine", "manhattan",
"hamming"}

\item{n.trees}{More trees gives higher precision when querying}
}
\value{
RccAnny::Annoy\link{metric} class
}
\description{
Build the annoy index
Ref: Seurat
}
