% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\docType{data}
\name{ArchRPalettes}
\alias{ArchRPalettes}
\title{List of color palettes that can be used in plots}
\format{
An object of class \code{list} of length 30.
}
\usage{
ArchRPalettes
}
\description{
A collection of some original and some borrowed color palettes to
provide appealing color aesthetics for plots in ArchR.
}
\details{
Comments in ArchR:
DISCLOSURE: This is a collection of palettes that includes some
original palettes and some palettes originally
implemented by others in other packages.
They are included here for convenience because they help improve
plot aesthetics.
NOTE: all palettes included in the "Primarily Continuous Palettes"
section should also work for discrete usage but not vice versa.
Each continuous palette has been ordered by color to generate a
visually appealing discrete palette.
}
\keyword{datasets}
