% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateATACWithRNA.R
\name{calOvlpScoreOld}
\alias{calOvlpScoreOld}
\title{Calculate overlap scores.
Old function, not expored and deprecated}
\usage{
calOvlpScoreOld(t1, t2)
}
\description{
Calculate overlap scores.
Old function, not expored and deprecated
}
