% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{checkArgsExistOrStop}
\alias{checkArgsExistOrStop}
\title{Check if element in argument has no defination
argument generated from package optparse
But this function can be used for a list}
\usage{
checkArgsExistOrStop(args)
}
\arguments{
\item{args}{list-like object with attribute names}
}
\value{
None, if no defined attribute in args, stop will be called
}
\description{
Check if element in argument has no defination
argument generated from package optparse
But this function can be used for a list
}
