% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{checkFileExistOrStop}
\alias{checkFileExistOrStop}
\title{Check if file/dir exist, if not, then stop}
\usage{
checkFileExistOrStop(f)
}
\arguments{
\item{f}{characters, a file or director}
}
\value{
None, will send stop signal if
f doesn't exist.
}
\description{
Check if file/dir exist, if not, then stop
}
