% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{cleanOutfile}
\alias{cleanOutfile}
\title{Prepare path for a new output file.}
\usage{
cleanOutfile(outf, remove = TRUE)
}
\arguments{
\item{outf}{characters, file name with desired path}

\item{remove}{bool, if remove old file, default TRUE}
}
\value{
None
}
\description{
Prepare path for a new output file.
}
