% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnapATACUtils.R
\name{downSampleOnSnap}
\alias{downSampleOnSnap}
\title{Down sample snap object based on a given cluster meta.}
\usage{
downSampleOnSnap(snap, cluster, n = 200)
}
\arguments{
\item{snap}{SnapObject from SnapATAC}

\item{cluster}{vector of characters/integer
aligned with the cells in snap in order.}

\item{n}{integer, number of samples in each unique cluster id.}
}
\value{
SnapObject
}
\description{
Down sample snap object based on a given cluster meta.
}
