% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnapATACUtils.R
\name{emptySnapGraph}
\alias{emptySnapGraph}
\title{Empty snap's KNN graph if it has.}
\usage{
emptySnapGraph(s)
}
\arguments{
\item{s}{snap object}
}
\value{
snap object
}
\description{
Empty snap's KNN graph if it has.
}
