% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnapATACUtils.R
\name{filterSnapBmat}
\alias{filterSnapBmat}
\title{Filter bmat
removing common bins and those overlapped with blacklist}
\usage{
filterSnapBmat(
  snap = NULL,
  snapFile = NULL,
  low = 0,
  high = 0.99,
  blackListFile = NULL,
  blacklistGR = NULL,
  excludeChr = "random|chrM",
  addBmatIfEmpty = FALSE,
  binSize = 5000,
  ncores = 1
)
}
\arguments{
\item{snap}{snap object of SnapATAC, default NULL}

\item{snapFile}{characters name of the snap File to load, default NULL}

\item{low}{double threshold, default 0 (bin freq <= low will be removed)}

\item{high}{double threshold, default 0.99 (bin freq > high will be removed)}

\item{blackListFile}{characters file of the blacklistGR, default NULL}

\item{blacklistGR}{GenomicRanges object, default NULL}

\item{excludeChr}{characters used to filter bmat, default "random|chrM"}

\item{addBmatIfEmpty}{bool, default FALSE}

\item{binSize}{integer dim of bmat, default 5,000}

\item{ncores}{integer number of cores to use for parallel, default 1}
}
\value{
SnapObject
}
\description{
Filter bmat
removing common bins and those overlapped with blacklist
}
