% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{getColorPalFn}
\alias{getColorPalFn}
\title{Get color palatte function using viridis_pal fun from viridis package.}
\usage{
getColorPalFn(
  types,
  b = 0,
  e = 1,
  direction = 1,
  option = "D",
  showColor = FALSE
)
}
\arguments{
\item{types}{vector/array, which we want to assign each element with a distinct color}

\item{b}{numeric, \link{0,1} color begin value, 0.05 default}

\item{e}{numeric, \link{0,1} color end value, 0.2 default, no less than b
NOTE, both b and e are calculated from the right-side in the color scales.}

\item{direction}{integer default is 1}

\item{option}{color palette to use, default is "D"
NOTE: "magma" (or "A"); "inferno" (or "B");"plasma" (or "C");
"viridis" (or "D") ;"cividis" (or "E");"rocket" (or "F");
"mako" (or "G"); "turbo" (or "H")}

\item{showColor}{bool, after generating the color
if need to show colors with scales::show_col, default FALSE}
}
\value{
color palatte function, f(n) will get n different colors.
}
\description{
See details:
https://cran.r-project.org/web/packages/viridis/vignettes/intro-to-viridis.html
}
