% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{getContinuousColors}
\alias{getContinuousColors}
\title{Get continous color with linear interplotation}
\usage{
getContinuousColors(b = 0, e = 1, n = 50, pal = "Zissou1", type = "continuous")
}
\arguments{
\item{b}{numeric, region start, 0 as default}

\item{e}{numeric, region end, 1 as default}

\item{n}{integer, number of intervals for explicit color points withn \link{b, e}}

\item{pal}{characters, color palette, "Zissou1" as default}

\item{type}{characters, parameter for wesanderson pacakge, "continuous" as default}
}
\value{
function, f(value \in \link{b, e}) will return a color
}
\description{
Use circlize::colorRamp2 to do the linear interplotation
Use wasanderson::was_palette to generate the explict colors needed,
which could be replaced in the future with other more common packages.
}
