% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmat.R
\name{getCutoffFromShuffledBmat}
\alias{getCutoffFromShuffledBmat}
\title{Get enriched bin's percentage cutoff based on shuffledBmat.}
\usage{
getCutoffFromShuffledBmat(sBmat, scale = 3, noLessThan = 0.04)
}
\arguments{
\item{sBmat}{sparseMatrix, cell by bins, randomly shuffled.}

\item{scale}{double, default is 3.0, scale the standard deviation.}

\item{noLessThan}{double, default is 0.04.
cutoff should be no less than this.}
}
\value{
double
}
\description{
Get enriched bin's percentage cutoff based on shuffledBmat.
}
