% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmat.R
\name{getIsolatedGroup}
\alias{getIsolatedGroup}
\title{Based on pval matrix, link the nodes without significant p-values.
This is achieved by recursive way, which I think is too difficult.
Should have much simpler method.}
\usage{
getIsolatedGroup(pvalmat, pthres = 0.1)
}
\arguments{
\item{pvalmat}{symmetric matrix}

\item{pthres}{double, default is 0.1}
}
\value{
list of vectors, each vector is a graph module
defined by the pval matrix.
}
\description{
Based on pval matrix, link the nodes without significant p-values.
This is achieved by recursive way, which I think is too difficult.
Should have much simpler method.
}
