% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe.R
\name{getMatchDataFrame}
\alias{getMatchDataFrame}
\title{Select subset of rows of data.frame based on
the order of defined row keys.
Keys not in the data.frame will be ignored.}
\usage{
getMatchDataFrame(df, rowKeys, selectKeys)
}
\arguments{
\item{df}{data.frame}

\item{rowKeys}{vector, row keys for df}

\item{selectKeys}{vector, keys we want to select in order}
}
\value{
data.frame
}
\description{
Select subset of rows of data.frame based on
the order of defined row keys.
Keys not in the data.frame will be ignored.
}
