% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{getMultipleColors}
\alias{getMultipleColors}
\title{Get multiple distinct colors using viridis package.}
\usage{
getMultipleColors(types, b = 0.05, e = 0.2, pal = turbo, showColor = FALSE)
}
\arguments{
\item{types}{vector/array, which we want to assign each element with a distinct color}

\item{b}{numeric, \link{0,1} color begin value, 0.05 default}

\item{e}{numeric, \link{0,1} color end value, 0.2 default, no less than b
NOTE, both b and e are calculated from the right-side in the color scales.}

\item{pal}{literal, color palette to use, default is turbo (no "" needed)}

\item{showColor}{bool, after generating the color
if need to show colors with scales::show_col, default FALSE}
}
\value{
vector of characters with types as names. Each element represents one color.
}
\description{
See details:
https://cran.r-project.org/web/packages/viridis/vignettes/intro-to-viridis.html
}
