% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateATACWithRNA.R
\name{getOverlapMatrix}
\alias{getOverlapMatrix}
\title{Calculate Overlap Score matrix for ATAC and RNA jointly embedding metaTable}
\usage{
getOverlapMatrix(
  meta,
  ident = "coembed.idents",
  atacCol = "MajorType",
  rnaCol = "ClusterName"
)
}
\arguments{
\item{meta}{data.frame,
three columns as ident, atacCluster, rnaCluster defined as in the following
parameters}

\item{ident}{characters, name of column ident, i.e. the cluster Ids for
the co-embedded seurat, "coembed.idents" as default.}

\item{atacCluster}{characters, name of column atac,
"MajorType"as default}

\item{rnaCluster}{characters, name of column rna,
"ClusterName" as default}
}
\value{
data.frame of numeric, atac by rna
}
\description{
Calculate Overlap Score matrix for ATAC and RNA jointly embedding metaTable
}
