% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmat.R
\name{getPvalueOfNDiff.default}
\alias{getPvalueOfNDiff.default}
\title{Get number of diff bins and the pvalue given two groups.}
\usage{
getPvalueOfNDiff.default(
  mat,
  group,
  ith = 1,
  jth = 2,
  threshold = 0.04,
  nperm = 20
)
}
\arguments{
\item{group}{vector of integer, cluster index of cells}

\item{ith}{integer, default 1}

\item{jth}{integer, default 2}

\item{threshold}{double, default 0.04}

\item{nperm}{integer, default 20}

\item{bmat}{sparseMatrix, cell by bin}
}
\value{
vector
first is number of diff bin, second is p-value
}
\description{
Get number of diff bins and the pvalue given two groups.
}
