% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmat.R
\name{getPvalueOfNDiff.multiGroup}
\alias{getPvalueOfNDiff.multiGroup}
\title{Get number of diff bins and the pvalue for any two groups}
\usage{
getPvalueOfNDiff.multiGroup(
  mat,
  group,
  uniqueGroupWithName = NULL,
  threshold = 0.04,
  nperm = 20
)
}
\arguments{
\item{group}{vector of integer, cluster index of cells}

\item{uniqueGroupWithName}{vector, default NULL
unique names for group}

\item{threshold}{double, default 0.04}

\item{nperm}{integer, default 20}

\item{bmat}{sparseMatrix, cell by bin}
}
\value{
list of symmetric matrix
"diffbin" are number of diff bin, "p" are p-values.
each matrix has the names defined by names of uniqueGroupWithName
generated by this function.
}
\description{
Get number of diff bins and the pvalue for any two groups
}
