% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn.R
\name{\%||\%}
\alias{\%||\%}
\title{OR of NULL.
Ref: Seurat
If the first element is not NULL, then return the first one,
else return the second one.}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{R object}

\item{rhs}{R object}
}
\value{
lhs if lhs is not NULL, else rhs
}
\description{
OR of NULL.
Ref: Seurat
If the first element is not NULL, then return the first one,
else return the second one.
}
