% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmat.R
\name{igraphModule2List}
\alias{igraphModule2List}
\title{Transfer modules from igraph::components to a list.}
\usage{
igraphModule2List(modules, prefix = "c")
}
\arguments{
\item{modules}{list, returned from igraph::components}

\item{prefix}{characters, default is "c"}
}
\value{
list of vector / scalar with names
}
\description{
Transfer modules from igraph::components to a list.
}
