% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnapATACUtils.R
\name{landmarkEmbedding}
\alias{landmarkEmbedding}
\title{SnapATAC landmark embedding.}
\usage{
landmarkEmbedding(
  snap = NULL,
  snapFile = NULL,
  blacklistGR = NULL,
  blackListFile = NULL,
  binSize = 5000,
  nPC = 50,
  ncores = 1,
  outFile = NULL,
  excludeChr = "random|chrM",
  removeBmat = FALSE,
  removeJmat = FALSE
)
}
\arguments{
\item{snap}{snap object of SnapATAC, default NULL}

\item{snapFile}{characters name of the snap File to load, default NULL}

\item{blacklistGR}{GenomicRanges object, default NULL}

\item{blackListFile}{characters file of the blacklistGR, default NULL}

\item{binSize}{integer dim of bmat, default 5,000}

\item{nPC}{integer dim of embeding for embedding, default 50}

\item{ncores}{integer number of cores to use for parallel, default 1}

\item{outFile}{characters to which we write snap with embedding, default NULL}

\item{excludeChr}{characters used to filter bmat, default "random|chrM"}

\item{removeBmat}{bool default FALSE}

\item{removeJmat}{bool default FALSE}
}
\value{
SnapObject
}
\description{
SnapATAC landmark embedding.
}
