% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmat.R
\name{listOfNumeric2str}
\alias{listOfNumeric2str}
\title{Transer list of numeric to string}
\usage{
listOfNumeric2str(module, sepkv = ":", sepv = ".", sepk = ",")
}
\arguments{
\item{module}{list of numeric}

\item{sepkv}{characters, sep between key and value, default is ":"}

\item{sepv}{characters, sep between values, default is "."}

\item{sepk}{characters, sep between keys, default is ","}
}
\value{
characters
}
\description{
Transer list of numeric to string
}
