% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{loadRData}
\alias{loadRData}
\title{load RData file like readRDS}
\usage{
loadRData(fileRData, pattern = "\\\\.RData", check = TRUE)
}
\arguments{
\item{fileRData}{characters, RData file.}

\item{pattern}{characters, to check a RData file, default is "\\.RData"}

\item{check}{bool, if we need to check fileRData ends with .RData, default TRUE}
}
\value{
R object
}
\description{
load RData file like readRDS
}
