% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnapATACUtils.R
\name{orderSnap}
\alias{orderSnap}
\title{Order Snap based on the sample information
This is critical for adding bmat and other related functions in SnapATAC}
\usage{
orderSnap(s)
}
\arguments{
\item{s}{snap object}
}
\value{
snap object
}
\description{
Order Snap based on the sample information
This is critical for adding bmat and other related functions in SnapATAC
}
