% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnapATACUtils.R
\name{plot2D}
\alias{plot2D}
\title{Generate list of ggplot figures of umap/tsne on the features
If some feature is not in the meta, will have a NULL.}
\usage{
plot2D(
  embed,
  meta,
  names = c("cluster", "tsse", "log10UMI"),
  discretes = c(TRUE, FALSE, FALSE),
  legends = c(FALSE, TRUE, TRUE),
  addLabels = c(TRUE, FALSE, FALSE),
  n = 10000,
  ...
)
}
\arguments{
\item{embed}{data.frame cell by (x, y) dim, rownames are cells.}

\item{meta}{data.frame cell by features, rownames are cells,
colnames are features.}

\item{names}{character vector, features to be draw
Default is c("cluster", "tsse", "logumi")}

\item{discretes}{bool vector, default is c(T, F, F)}

\item{legends}{bool vector, default is c(F, T, T).}

\item{addLabels}{bool vectors, default is c(T, F, F).}

\item{n}{integer, number of down samples, default 10,000}

\item{...}{parameters ggPoint, but seems no use there.}
}
\value{
list of ggplot object
}
\description{
Generate list of ggplot figures of umap/tsne on the features
If some feature is not in the meta, will have a NULL.
}
