% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnapATACUtils.R
\name{plotConsensus}
\alias{plotConsensus}
\title{Consensus plot.}
\usage{
plotConsensus(consensusFile, type, outPDF, width = 7, height = 7)
}
\arguments{
\item{consensusFile}{characters}

\item{type}{characters, cell type name for title}

\item{outPDF}{characters}

\item{width}{numeric, default is 7}

\item{height}{numeric, default is 7}
}
\value{
None.
Side effect: generate the figure file.
}
\description{
Consensus plot.
}
