% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure.R
\name{plotFeatureSingle.SnapATAC}
\alias{plotFeatureSingle.SnapATAC}
\title{plotFeatureSingle from SnapATAC package
Allow the object can be a matrix of umap/tsne or a snap}
\usage{
plotFeatureSingle.SnapATAC(
  snap = NULL,
  featureValue,
  embedmat = NULL,
  method = c("tsne", "umap"),
  pointSize = 0.1,
  pointShape = 19,
  downSample = 10000,
  pdfile = NULL,
  pdfWidth = 7,
  pdfHeight = 7,
  quantiles = c(0.01, 0.99),
  fontLab = 2,
  ...
)
}
\arguments{
\item{snap}{A snap object, default is NULL}

\item{featureValue}{Feature enrichment value for each cell.
Value will be normalized betweeen 0 and 1.}

\item{embedmat}{a matrix of umap or tsne, default is NULL}

\item{method}{Visulization method c("tsne", "umap").}

\item{pointSize}{Point size \link{1}.}

\item{pointShape}{Point shape type \link{19}.}

\item{downSample}{Downsample the original cells to
down.sample cells to ovoid large dataset \link{10,000}.}

\item{pdfile}{pdf file name to save the plot \link{NULL}.}

\item{pdfWidth}{Width of the graphics region in inches \link{7}.}

\item{pdfHeight}{Height of the graphics region in inches \link{7}.}

\item{quantiles}{Feature value outside this range will be removed \link{c(0.01, 0.99)}}

\item{...}{Arguments passed to plot method.}

\item{fontLabl}{integer default is 2}
}
\description{
plotFeatureSingle from SnapATAC package
Allow the object can be a matrix of umap/tsne or a snap
}
