% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{prepareOutdir}
\alias{prepareOutdir}
\title{If dir does not exists, then create it.}
\usage{
prepareOutdir(d)
}
\arguments{
\item{d}{characters, dir name}
}
\value{
None
}
\description{
If dir does not exists, then create it.
}
