% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{prepareOutfile}
\alias{prepareOutfile}
\title{If file does not exists, then prepare it.}
\usage{
prepareOutfile(f)
}
\arguments{
\item{f}{characters, file name}
}
\value{
None
}
\description{
If file does not exists, then prepare it.
}
