% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnapATACUtils.R
\name{queryEmbedding}
\alias{queryEmbedding}
\title{SnapATAC query embedding.}
\usage{
queryEmbedding(
  snapLandmark = NULL,
  snapLandmarkFile = NULL,
  snapQuery = NULL,
  snapQueryFile = NULL,
  binSize = 5000,
  outFile = NULL,
  removeBmat = TRUE,
  removeJmat = TRUE
)
}
\arguments{
\item{snapLandmark}{snap object of SnapATAC, default NULL}

\item{snapQuery}{snap object of SnapATAC, default NULL}

\item{snapQueryFile}{characters name of the snap File to load, default NULL}

\item{binSize}{integer dim of bmat, default 5,000}

\item{outFile}{characters to which we write snap with embedding, default NULL}

\item{removeBmat}{bool default FALSE}

\item{removeJmat}{bool default FALSE}

\item{snaplandmarkFile}{characters name of the snap File to load, default NULL}
}
\value{
SnapObject
}
\description{
SnapATAC query embedding.
}
