% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SnapATACUtils.R
\name{runLeiden}
\alias{runLeiden}
\title{Run Leiden Algorithm.}
\usage{
runLeiden(
  snap = NULL,
  snapFile = NULL,
  r = 0.5,
  pt = "RB",
  seed = 10,
  pathToPython = NULL,
  outLeidenFile = NULL,
  outClusterMetaCSV = NULL,
  outClusterPDF = NULL,
  pdfn = NULL,
  colName = "cluster",
  dims = 1:30,
  umapNcores = 1,
  ...
)
}
\arguments{
\item{snap}{SnapObject default NULL}

\item{snapFile}{characters default NULL}

\item{r}{numeric resolution paramter for leiden, default 0.5}

\item{pt}{characters partition type for leiden, default "RB"}

\item{seed}{integer default NULL}

\item{pathToPython}{characters where the python is, default NULL}

\item{outLeidenFile}{characters default NULL}

\item{outClusterMetaCSV}{characters default NULL}

\item{outClusterPDF}{characters default NULL}

\item{pdfn}{function used to draw figures for outClusterPDF
the inputs are snap object and others
default NULL.}

\item{colName}{characters used for record cluster namae in snap meta,
default is "cluster". if snap@metaData has this column,
will use \link{colName}.1 instead}

\item{dims}{integer or vector for UMAP default is 1:30}

\item{umapNcores}{integer default 1}

\item{...}{use for pdfn function}
}
\value{
SnapObject, slot "cluster" is the result
}
\description{
Run Leiden Algorithm.
}
