% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{sampleBasedOnDepth2}
\alias{sampleBasedOnDepth2}
\title{Sample index based on sequence depth defined by log10UMI
without repeat.}
\usage{
sampleBasedOnDepth2(log10UMI, n, seed = 2022)
}
\arguments{
\item{log10UMI}{numeric vector}

\item{n}{integer, number of samples needed}

\item{seed}{integer, random seed, 2022 (default)}
}
\value{
vector of integer
}
\description{
Sample index based on sequence depth defined by log10UMI
without repeat.
}
