% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmat.R
\name{shuffleBmat}
\alias{shuffleBmat}
\title{Shuffle bmat in both row and column.}
\usage{
shuffleBmat(bmat)
}
\arguments{
\item{bmat}{sparseMatrix, cell by bins}
}
\value{
sparseMatrix
}
\description{
Shuffle bmat in both row and column.
}
