% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmat.R
\name{strToAdjmat}
\alias{strToAdjmat}
\title{Transfer specific strings to adjacent matrix
A typical string like: pval@c1-c2:0,c1-c3:0}
\usage{
strToAdjmat(
  mystring,
  title = "pval@",
  sepg = ",",
  sepk = "-",
  sepkv = ":",
  clusterPrefix = "c"
)
}
\arguments{
\item{mystring}{characters, like: pval@c1-c2:0,c1-c3:0}

\item{title}{characters, default is "pval@"}

\item{sepg}{characters, sep between groups, default is ","}

\item{sepk}{characters, sep between keys, default is "-"}

\item{sepkv}{characters, sep between key and value, default is ":"}

\item{clusterPrefix}{characters, default is "c"}
}
\value{
Matrix symmetric
}
\description{
Transfer specific strings to adjacent matrix
A typical string like: pval@c1-c2:0,c1-c3:0
}
