% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmat.R
\name{uptri2str}
\alias{uptri2str}
\title{Transfer the upper tri of a matrix to a string}
\usage{
uptri2str(mat, sepkv = ":", sepk = "-", sepg = ",", digit = 2)
}
\arguments{
\item{mat}{Matrix}

\item{sepkv}{characters, sep between key and value, default is ":"}

\item{sepk}{characters, sep between keys, default is "-"}

\item{sepg}{characters, sep between groups, default is ","}

\item{digit}{integer, default 2}
}
\value{
characters
}
\description{
Transfer the upper tri of a matrix to a string
}
