% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{getMultipleColors}
\alias{getMultipleColors}
\title{Get multiple distinct colors using viridis package.}
\usage{
getMultipleColors(
  types,
  alpha = 1,
  b = 0,
  e = 1,
  direction = 1,
  pal = viridis,
  showColor = FALSE
)
}
\arguments{
\item{types}{vector/array, which we want to assign each element with a distinct color}

\item{alpha}{numeric, default is 1}

\item{b}{numeric, \link{0,1} color begin value, default is 0}

\item{e}{numeric, \link{0,1} color end value, default is 1 , no less than b
NOTE, both b and e are calculated from the right-side in the color scales.}

\item{direction}{1 or -1, default is 1
If 1, colors are ordered from darkest to lightest.
If -1, the order of colors is reversed.}

\item{pal}{literal, color palette to use, default is viridis (no "" needed)}

\item{showColor}{bool, after generating the color
if need to show colors with scales::show_col, default FALSE}
}
\value{
vector of characters with types as names. Each element represents one color.
}
\description{
See details:
https://cran.r-project.org/web/packages/viridis/vignettes/intro-to-viridis.html
}
