% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure.R
\name{plotDotplot}
\alias{plotDotplot}
\title{Get Seurat dot plot for snap gmat.}
\usage{
plotDotplot(
  snap = NULL,
  snapList = NULL,
  cluster = NULL,
  features,
  groupBy = "cluster",
  cellidSep = ".",
  axis.text.x.angle = 30,
  axis.text.x.hjust = 1,
  ...
)
}
\arguments{
\item{snap}{snap object, default is NULL}

\item{snapList}{list of snap object, default is NULL}

\item{cluster}{vector of characters, integers, or factors
must have names, which are the cell ids. Cluster size can be
larger than snap or snapList size, but have to make sure all
the cells in snap or snapList have the cluster identification.}

\item{features}{vector of characters, genes for plotting.}

\item{cellidSep}{characters default is "."}

\item{axis.text.x.angle}{numeric, default is 30}

\item{axis.text.x.hjust}{numeric, default is 1}

\item{...}{used for Seurat::DotPlot function}

\item{group.by}{characters default is cluster}
}
\value{
ggplot object
}
\description{
Get Seurat dot plot for snap gmat.
}
