% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure.R
\name{plotFeatureSingle}
\alias{plotFeatureSingle}
\title{plotFeatureSingle
Allow the object can be a matrix of umap/tsne or a snap}
\usage{
plotFeatureSingle(
  snap = NULL,
  featureValue,
  embedmat = NULL,
  method = c("tsne", "umap"),
  downSample = 10000,
  pdfile = NULL,
  pdfWidth = 7,
  pdfHeight = 7,
  quantiles = c(0.01, 0.99),
  pal = viridis::viridis(n = 256),
  colorTitle = "gene score",
  pointSize = 0.1,
  labelSize = 5,
  baseSize = 12,
  legendSize = 8,
  discrete = FALSE,
  labelMeans = FALSE,
  ...
)
}
\arguments{
\item{snap}{A snap object, default is NULL}

\item{featureValue}{Feature enrichment value for each cell.
Value will be normalized betweeen 0 and 1.}

\item{embedmat}{a matrix of umap or tsne, default is NULL}

\item{method}{Visulization method c("tsne", "umap").}

\item{downSample}{Downsample the original cells to
down.sample cells to ovoid large dataset \link{10,000}.}

\item{pdfile}{pdf file name to save the plot \link{NULL}.}

\item{pdfWidth}{Width of the graphics region in inches \link{7}.}

\item{pdfHeight}{Height of the graphics region in inches \link{7}.}

\item{quantiles}{Feature value outside this range will be removed \link{c(0.01, 0.99)}}

\item{pal}{colors for ploting continuous numbers, default viridis::viridis(n=256)}

\item{colorTitle}{legend title, default is "gene score"}

\item{pointSize}{numeric, default is 0.1}

\item{labelSize}{integer default is 5}

\item{baseSize}{integer default is 12}

\item{legendSize}{integer default is 8}

\item{...}{Arguments passed to ggPoint method}
}
\value{
ggplot object
}
\description{
plotFeatureSingle
Allow the object can be a matrix of umap/tsne or a snap
}
